Add support for negative rates on barrier options.
Instead of flooring to zero as done in the pricing module, here we need
a very small value to avoid similar divide by 0 problems. For this,
we use the "SMALL_VALUE_MU" which seems tailor-made for the purpose.

Also, some cleanup and rationalization of the test file to make it
easier to understand both the source and output. Plus, of course, adding
a couple test cases for the improved functions.

All of the values seem reasonable to the untrained eye.